+++ src/enumeration/dependency-type.ts
@@ -1,0 +1,3 @@
+export enum DependencyType {
+    VUE=1, FILE=2
+}