+++ src/index.ts
@@ -1,5 +1,6 @@
 export * from './abstraction';
 export * from './classification';
+export * from './enumeration';
 export * from './function';
 export * from './globalization';
 export * from './loader';